requirements.txt
streamlit==1.28.1
moviepy==1.0.3
gtts==2.4.0
edge-tts==6.1.8
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
aiofiles==23.2.0
ffmpeg-python==0.2.0
openai>=1.0.0
pydub==0.25.1
numpy>=1.21.0
streamlit-option-menu==0.3.2
streamlit-tags==1.2.7
python-multipart==0.0.6
watchdog==3.0.0
decorator==4.4.2
imageio==2.31.1
imageio-ffmpeg==0.4.8
proglog==0.1.10
opencv-python==4.8.1.78
setuptools==80.0.0  # Pin to version before removal
tqdm>=4.66.0
